Show range selection in days selection controller
