+++ Habit-Calendar/Habit-Calendar/Views/Calendar/CalendarDayCell.swift
@@ -11,6 +11,13 @@ import JTAppleCalendar
 /// Cell in charge of displaying the calendar's days.
 @IBDesignable class CalendarDayCell: JTAppleCell {
 
+    // MARK: Types
+
+    /// The kind of challenge day being displayed within the challenge's range of days.
+    enum RangePosition {
+        case begin, inBetween, end, none
+    }
+
     // MARK: Parameters
 
     /// The text's default color.
@@ -35,12 +42,54 @@ import JTAppleCalendar
         return circleView
     }()
 
-    /// The cell's bottom separator.
-    lazy var bottomSeparator: CALayer = {
-        let separatorLayer = CALayer()
-        separatorLayer.backgroundColor = UIColor(red: 236/255, green: 240/255, blue: 241/255, alpha: 1).cgColor
+    /// The day's position within the challenge's days.
+    /// Default value is none.
+    var position: RangePosition = .none {
+        didSet {
+            setNeedsLayout()
+        }
+    }
 
-        return separatorLayer
+    /// The background view displaying the day's position in the challenge.
+    private(set) lazy var rangeBackgroundView: UIView = {
+        let rangeView = UIView()
+        rangeView.translatesAutoresizingMaskIntoConstraints = false
+
+        return rangeView
+    }()
+
+    /// The range view's vertical constraint applied to all situations.
+    private lazy var verticalConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.centerYAnchor.constraint(equalTo: dayTitleLabel.centerYAnchor)
+    }()
+
+    /// The range view's height constraint applied to all situations.
+    private lazy var heightConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.heightAnchor.constraint(
+            equalTo: dayTitleLabel.heightAnchor,
+            multiplier: 1,
+            constant: 10
+        )
+    }()
+
+    /// The range view's width constraint applied to all situations.
+    private lazy var widthConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 2, constant: 0)
+    }()
+
+    /// The range view's begin horizontal constraint.
+    private lazy var beginHorizontalConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.leadingAnchor.constraint(equalTo: dayTitleLabel.leadingAnchor, constant: -10)
+    }()
+
+    /// The range view's inBetween horizontal constraint.
+    private lazy var inBetweenHorizontalConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.centerXAnchor.constraint(equalTo: contentView.centerXAnchor)
+    }()
+
+    /// The range view's end horizontal constraint.
+    private lazy var endHorizontalConstraint: NSLayoutConstraint = {
+        return rangeBackgroundView.trailingAnchor.constraint(equalTo: dayTitleLabel.trailingAnchor, constant: 10)
     }()
 
     /// MARK: Life cycle
@@ -50,6 +99,13 @@ import JTAppleCalendar
 
         handleSubviews()
         applyLayout()
+
+        switch position {
+        case .begin, .inBetween, .end:
+            rangeBackgroundView.layer.cornerRadius = (dayTitleLabel.intrinsicContentSize.height + 10) / 2
+        default:
+            rangeBackgroundView.layer.cornerRadius = 0
+        }
     }
 
     override func prepareForInterfaceBuilder() {
@@ -64,6 +120,11 @@ import JTAppleCalendar
         backgroundColor = .white
         circleView.backgroundColor = .clear
         dayTitleLabel.textColor = defaultTextColor
+
+        // Deactivate all horizontal constraints (specific constraints for each position type).
+        beginHorizontalConstraint.isActive = false
+        inBetweenHorizontalConstraint.isActive = false
+        endHorizontalConstraint.isActive = false
     }
 
     // MARK: Imperatives
@@ -72,13 +133,12 @@ import JTAppleCalendar
     func handleSubviews() {
         guard !contentView.subviews.contains(circleView),
             !contentView.subviews.contains(dayTitleLabel),
-            bottomSeparator.superlayer == nil else {
+            !contentView.subviews.contains(rangeBackgroundView) else {
                 return
         }
         contentView.addSubview(circleView)
         contentView.addSubview(dayTitleLabel)
-        contentView.layer.addSublayer(bottomSeparator)
-
+        contentView.addSubview(rangeBackgroundView)
     }
 
     /// Applies the layout to the subviews, if appropriate.
@@ -92,11 +152,29 @@ import JTAppleCalendar
         circleView.heightAnchor.constraint(equalToConstant: 35).isActive = true
         circleView.layer.cornerRadius = 35 * 0.5
 
-        bottomSeparator.frame = CGRect(
-            x: 0,
-            y: contentView.frame.size.height - 1,
-            width: contentView.frame.size.width,
-            height: 1
-        )
+        // The circle view should always be in the front.
+        contentView.bringSubviewToFront(circleView)
+        contentView.bringSubviewToFront(dayTitleLabel)
+
+        rangeBackgroundView.isHidden = false
+
+        verticalConstraint.isActive = true
+        heightConstraint.isActive = true
+        widthConstraint.isActive = true
+
+        // Apply the rangeBgView's layout, according to the current position.
+        switch position {
+        case .begin:
+            beginHorizontalConstraint.isActive = true
+
+        case .inBetween:
+            inBetweenHorizontalConstraint.isActive = true
+
+        case .end:
+            endHorizontalConstraint.isActive = true
+
+        case .none:
+            rangeBackgroundView.isHidden = true
+        }
     }
 }
